// src/interfaces/Prices.interface.tsx
// Generated by https://quicktype.io

export interface PricesResponse {
  prices: Prices;
}

export interface Prices {
  oneandhalfkgPrice: number;
  twokgPrice: number;
  threekgPrice: number;
  fourkgPrice: number;
  fivekgPrice: number;
  sixkgPrice: number;
  sevenkgPrice: number;
  eightkgPrice: number;
  ninekgPrice: number;
  tenkgPrice: number;

  elevenkgPrice: number;
  twelvekgPrice: number;
  thirteenkgPrice: number;
  fourteenkgPrice: number;
  fifteenkgPrice: number;
  sixteenkgPrice: number;
  seventeenkgPrice: number;
  eighteenkgPrice: number;
  nineteenkgPrice: number;

  twentykgPrice: number;
  twentyonekgPrice: number;
  twentytwokgPrice: number;
  twentythreekgPrice: number;
  twentyfourkgPrice: number;
  twentyfivekgPrice: number;
  twentysixkgPrice: number;
  twentysevenkgPrice: number;
  twentyeightkgPrice: number;
  twentyninekgPrice: number;

  thirtykgPrice: number;
  thirtyonekgPrice: number;
  thirtytwokgPrice: number;
  thirtythreekgPrice: number;
  thirtyfourkgPrice: number;
  thirtyfivekgPrice: number;
  thirtysixkgPrice: number;
  thirtysevenkgPrice: number;
  thirtyeightkgPrice: number;
  thirtyninekgPrice: number;

  fortykgPrice: number;
  fortyonekgPrice: number;
  fortytwokgPrice: number;
  fortythreekgPrice: number;
  fortyfourkgPrice: number;
  fortyfivekgPrice: number;
  fortysixkgPrice: number;
  fortysevenkgPrice: number;
  fortyeightkgPrice: number;
  fortyninekgPrice: number;

  fiftykgPrice: number;
  fiftyonekgPrice: number;
  fiftytwokgPrice: number;
  fiftythreekgPrice: number;
  fiftyfourkgPrice: number;
  fiftyfivekgPrice: number;
  fiftysixkgPrice: number;
  fiftysevenkgPrice: number;
  fiftyeightkgPrice: number;
  fiftyninekgPrice: number;

  sixtykgPrice: number;
  sixtyonekgPrice: number;
  sixtytwokgPrice: number;
  sixtythreekgPrice: number;
  sixtyfourkgPrice: number;
  sixtyfivekgPrice: number;
  sixtysixkgPrice: number;
  sixtysevenkgPrice: number;
  sixtyeightkgPrice: number;
  sixtyninekgPrice: number;

  seventykgPrice: number;
  seventyonekgPrice: number;
  seventytwokgPrice: number;
  seventythreekgPrice: number;
  seventyfourkgPrice: number;
  seventyfivekgPrice: number;
  seventysixkgPrice: number;
  seventysevenkgPrice: number;
  seventyeightkgPrice: number;
  seventyninekgPrice: number;

  eightykgPrice: number;
  eightyonekgPrice: number;
  eightytwokgPrice: number;
  eightythreekgPrice: number;
  eightyfourkgPrice: number;
  eightyfivekgPrice: number;
  eightysixkgPrice: number;
  eightysevenkgPrice: number;
  eightyeightkgPrice: number;
  eightyninekgPrice: number;

  ninetykgPrice: number;
  ninetyonekgPrice: number;
  ninetytwokgPrice: number;
  ninetythreekgPrice: number;
  ninetyfourkgPrice: number;
  ninetyfivekgPrice: number;
  ninetysixkgPrice: number;
  ninetysevenkgPrice: number;
  ninetyeightkgPrice: number;
  ninetyninekgPrice: number;
  onehundredkgPrice: number;
  createdAt?: string;
  updatedAt?: string;
  id?: string;
}